requirements.txt
streamlit
langchain
langchain-openai
openai
python-dotenv
pandas
matplotlib
seaborn
chromadb
